⚡️ :: DataStorage 분리
