+++ data/src/main/java/team/aliens/data/_datasource/local/LocalAuthDataSource.kt
@@ -1,0 +1,6 @@
+package team.aliens.data._datasource.local
+
+interface LocalAuthDataSource {
+
+
+}
