+++ local/src/main/java/team/aliens/local/datastore/storage/AuthDataStorage.kt
@@ -26,20 +26,4 @@ interface AuthDataStorage {
     suspend fun updateAutoSignInOption(
         autoSignIn: Boolean,
     )
-
-    suspend fun findFeature(): Feature
-
-    suspend fun findMealFeatureEnabled(): Boolean
-
-    suspend fun findNoticeFeatureEnabled(): Boolean
-
-    suspend fun findPointServiceEnabled(): Boolean
-
-    suspend fun findStudyRoomServiceEnabled(): Boolean
-
-    suspend fun findRemainsServiceEnabled(): Boolean
-
-    suspend fun saveFeature(
-        feature: Feature,
-    )
 }
